plugins/boto3/requirements.txt
boto3 == 1.*
pyyaml == 5.*
